requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.1
pypdf>=4.0.0
lancedb>=0.4.0
langchain>=0.2.10
langchain-community>=0.2.10
langchain-google-genai>=0.1.0
langgraph>=0.1.0
google-generativeai>=0.5.2
google-api-core>=2.17.0
dspy>=2.5.0
typing-extensions>=4.9.0
